Created distributed config load functions, classifier module and a segmenter class.
A classifier module and class were created. Configuration loading has been moved into
classes for classifiers, featurisers, hash functions, etc. instead of the datas module.
A segmenter module and class were created to handle training and segmentation. A bug
having to do with getting the mask to get the non-control labels from a label array
was fixed. Added ability to use random search with classifiers. None of this is connected
to the main program yet as it still uses the old datas module to work.