Fixed use_gpu argument.
use_gpu is not a command line argument rather than a configuration parameter.
Also added not implemented error to GPU version of histogram extractor as it is not complete yet.
Import of GPU related modules was moved to GPU histogram extractor function to avoid errors when GPU hardware is not present.